[BED-121] Rename package
Took 27 seconds
