+++ src/main/java/net/alphalightning/bedwars/feedback/visual/impl/FakeBlockVisualization.java
@@ -3,8 +3,8 @@ package net.alphalightning.bedwars.feedback.visual.impl;
 import net.alphalightning.bedwars.feedback.visual.Visualization;
 import net.alphalightning.bedwars.feedback.visual.manager.VisualizationManager;
 import net.alphalightning.bedwars.setup.map.MapSetup;
-import net.alphalightning.bedwars.utils.BedUtil;
-import net.alphalightning.bedwars.utils.BlockUtil;
+import net.alphalightning.bedwars.util.BedUtil;
+import net.alphalightning.bedwars.util.BlockUtil;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.World;
