+++ src/main/java/net/alphalightning/bedwars/feedback/visual/impl/ValuableSpawnerVisualization.java
@@ -5,7 +5,7 @@ import net.alphalightning.bedwars.feedback.visual.Visualization;
 import net.alphalightning.bedwars.feedback.visual.manager.VisualizationManager;
 import net.alphalightning.bedwars.game.SpawnerType;
 import net.alphalightning.bedwars.setup.map.MapSetup;
-import net.alphalightning.bedwars.utils.LocationUtil;
+import net.alphalightning.bedwars.util.LocationUtil;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.Material;
