+++ src/main/java/net/alphalightning/bedwars/setup/map/stages/gamemap/TeamLootspawnerConfigurationStage.java
@@ -12,7 +12,7 @@ import net.alphalightning.bedwars.setup.map.stages.LocationConfiguration;
 import net.alphalightning.bedwars.setup.map.stages.Stage;
 import net.alphalightning.bedwars.setup.map.stages.TeamConfiguration;
 import net.alphalightning.bedwars.translation.NamedTranslationArgument;
-import net.alphalightning.bedwars.utils.BlockUtil;
+import net.alphalightning.bedwars.util.BlockUtil;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.TranslatableComponent;
 import org.bukkit.Location;
